image_service/requirements.txt
Django==4.0.3
djangorestframework
Pillow==9.2.0